Address style changes

+++ src/lib.rs
@@ -7,5 +7,5 @@ pub const EXPIRATION_OFFSET: u64 = 3600;
 pub fn get_current_time() -> u64 {
     let start = SystemTime::now();
     let since_the_epoch = start.duration_since(UNIX_EPOCH).expect("Time went backwards");
-    since_the_epoch.as_secs() as u64
+    since_the_epoch.as_secs()
 }
